Implement new count_ps metric for extended counters and timers
If user is receivering timer data and wants to know how many
timer metrics are receiverd per second then a new count_ps metric
is needed.

The old rate metric makes sense for counters because it answers
the question "how much the counter value was incremented per second".
For timers the rate instead is answering "how many time units
(for example milliseconds) was spent each second. This
is why we need the count_ps to answer the missing question.